requirements.txt
# Requirements for the ML Deployment Workshop
# File: requirements.txt

# Core ML libraries
torch>=1.12.0
torchvision>=0.13.0
scikit-learn>=1.1.0
numpy>=1.21.0
joblib>=1.1.0

# API framework
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5

# Image processing
Pillow>=8.3.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
requests>=2.25.0
python-dotenv>=0.19.0
pydantic>=1.8.0

# For Jupyter notebooks (optional)
jupyter>=1.0.0
ipywidgets>=7.6.0

# For monitoring and logging
loguru>=0.6.0

# For handling different image formats
opencv-python>=4.5.0
